Add isSigned method and tests
Summary: A token will throw if the signature is invalid, but not indicate that there was no signature. This could allow an attacker with a JWT to modify the data by simply stripping the signature and changing the algorithm in the header to 'none', which is no bueno.

Test Plan: Unit tests added

Differential Revision: https://phabricator.ericstern.com/D39