Decouple decoding and verification
In reference to the previous commit, it's preferable to store a key id in the JWT header to indicate the algorithm and key to be used during verification. Previously this was not possible since the header containing the verification information wasn't available until after the verification process! This allows the token to be decoded without verification and the data accessed in an explicitly-insecure mode to obtain (or derive) the key and algorithm to perform verification correctly.

While this could be integrated into the decode process, it would be exceptionally complicated: at minimum, it would require some sort of "KeyStore" interface and dependency, and would require some sort of field whitelisting to handle dynamic keys. This change allows a simpler implementation:

```
<?php
$raw_token = $some_jwt_string;
$JWT = JWT::decode($raw_token);
$unsafe = $JWT->getUnverifiedClaims();
list($alg, $key) = someLookupFunction($unsafe);
$JWT->verify($alg, $key);
$safe = $JWT->getClaims();
```